feat(next): split Server/Client components for app and talk pages\n\n- Make  a Server Component rendering  (client)\n- Make  a Server Component rendering  (client)\n- Extract  and  under \n\nCloses #5
